+++ glfw/wl_platform.h
@@ -167,8 +167,7 @@ typedef struct _GLFWwindowWayland
 
     struct zwp_idle_inhibitor_v1*          idleInhibitor;
 
-    // This is a hack to prevent auto-iconification on creation.
-    GLFWbool                    wasFullScreen;
+    GLFWbool                    fullscreened;
 
     struct {
         GLFWbool                           serverSide;